rename table dir
