+++ src/components/OracleQueryTable/OracleQueryTable.tsx
