+++ src/components/OracleQueryTable/ProposeCells.tsx
