+++ src/components/OracleQueryTable/Row.tsx
