+++ src/components/OracleQueryTable/SettledCells.tsx
