+++ src/components/OracleQueryTable/VerifyCells.tsx
