+++ src/components/OracleQueryTable/style.tsx
