+++ src/components/index.ts
@@ -15,4 +15,4 @@ export { Pagination } from "./Pagination";
 export { Panel } from "./Panel/Panel";
 export { PanelBase } from "./Panel/PanelBase";
 export { RadioDropdown } from "./RadioDropdown";
-export { OracleQueryTable } from "./Table/OracleQueryTable";
+export { OracleQueryTable } from "./OracleQueryTable/OracleQueryTable";
